Fix output issue. Now fully working VM!
This commit fixes an output issue. Basically, this console crate has a
bug where if you use 'read_char' the cursor does not reset after
printing a newline. This completely messes up the output.

The LC3 now uses a ridiculous hack of doing 'Term::clear_line' after
printing a newline which fixes the issue but is quite inefficient.